+++ python_dev_tools/formatters/pyupgrade.py
@@ -1,0 +1,10 @@
+"""Pyupgrade formatter management."""
+from .common import Formatter
+
+
+class PyupgradeFormatter(Formatter):
+    """Pyupgrade formatter management."""
+
+    name = "pyupgrade"
+    path = "pyupgrade"
+    cli_args = ["--py36-plus"]